requirements.txt
###### Requirements without Version Specifiers ######
streamlit-aggrid
openpyxl
nltk
pycountry
altair-catplot
xlrd
seaborn